# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super TTS daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/tts/super-tts-http.sock` (override with `SUPER_TTS_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/tts/super-tts-http.sock\" \\\n     -H \"Authorization: Bearer $TTS_TOKEN\" \\\n     http://tts.local/v1/ping\n```",
    "contact": {
      "name": "Super TTS",
      "url": "https://github.com/jorge-menjivar/super-tts"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://tts.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/registry/backend/list": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Browse the published backend catalog",
        "description": "Every backend published to the registry, with the models each serves and whether this machine can run it. `compatibility.compatible` is decided against the host's actual accelerators, so the list reflects what is installable here rather than what exists in general; pass `include_incompatible=true` to see the rest, each with a `compatibility.reason` worth showing the user.\n\n`update_available` is the daemon's answer rather than the client's arithmetic: it compares the installed `backend.toml` on disk against what the index offers, by `source`, and refuses a downgrade. An entry with no `installed_version` is an install, not an update.\n\nThis is what is *available*; `GET /backend/list` is what is installed. The index is cached and re-fetched on a schedule — `POST /registry/backend/refresh` forces that now.",
        "operationId": "list_registry_backends",
        "parameters": [
          {
            "name": "include_incompatible",
            "in": "query",
            "description": "Include entries this machine cannot run. Off by default, so the catalog\nshows what is actually installable here.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by transport — `wasm` or `subprocess`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "online",
            "in": "query",
            "description": "Filter by whether the backend calls out to a network service.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive substring match over name and description.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The catalog could not be fetched and nothing is cached (`registry_unavailable`). Retry, or force a fetch with `POST /registry/backend/refresh`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/registry/backend/refresh": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Re-fetch the backend catalog",
        "description": "Pulls the published index again rather than serving what is cached, and reports how many backends it now holds. Use it after a backend is published, or to clear an `index_stale` marker a listing came back with.\n\n`GET /registry/backend/list` refreshes on its own schedule; this forces it now. Idempotent, and concurrent calls coalesce into a single fetch, so a refresh button cannot start a stampede. The outcome is also published on the `registry_install` event topic, which is how a second client learns the catalog moved.",
        "operationId": "refresh_registry",
        "responses": {
          "200": {
            "description": "Refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The registry could not be reached (`registry_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Compatibility": {
        "type": "object",
        "required": [
          "compatible"
        ],
        "properties": {
          "compatible": {
            "type": "boolean"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "selected_asset": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SelectedAsset"
              }
            ]
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "speech_in_progress",
          "download_in_progress",
          "no_switch_in_progress",
          "model_not_loaded",
          "invalid_model",
          "invalid_backend",
          "cuda_unavailable",
          "invalid_device",
          "online_models_disabled",
          "invalid_audio_theme",
          "unsupported_language",
          "invalid_value",
          "not_found",
          "internal",
          "unknown"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-tts/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "IndexModel": {
        "type": "object",
        "description": "The browse-only model subset the catalog and host-compatibility filter need\nbefore download. The authoritative manifest (languages, files, …) ships as\nthe pinned `manifest` asset and is installed verbatim — it is not re-encoded\nhere. Also the leaf type for `/registry/backends` (`RegistryModel`).",
        "required": [
          "name",
          "supported_devices"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "supported_devices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "IndexOption": {
        "type": "object",
        "required": [
          "name",
          "label",
          "type"
        ],
        "properties": {
          "default": {},
          "label": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "IndexSecret": {
        "type": "object",
        "required": [
          "name",
          "label",
          "required"
        ],
        "properties": {
          "label": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          }
        }
      },
      "IndexStale": {
        "type": "object",
        "required": [
          "latest_attempted",
          "tag",
          "error",
          "since"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "latest_attempted": {
            "type": "string"
          },
          "since": {
            "type": "string"
          },
          "tag": {
            "type": "string"
          }
        }
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "backend_count"
        ],
        "properties": {
          "backend_count": {
            "type": "integer",
            "minimum": 0
          },
          "generated_at": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "RegistryBackend": {
        "type": "object",
        "required": [
          "id",
          "source",
          "version",
          "name",
          "license",
          "kind",
          "contract",
          "online",
          "supports_gpu",
          "supports_cpu",
          "models",
          "secrets",
          "options",
          "compatibility"
        ],
        "properties": {
          "allowed_hosts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "backend_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The backend's reverse-DNS identifier, or `None` when the registry entry\npredates it. Names the install directory."
          },
          "compatibility": {
            "$ref": "#/components/schemas/Compatibility"
          },
          "contract": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "index_stale": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IndexStale"
              }
            ]
          },
          "installed_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexModel"
            }
          },
          "name": {
            "type": "string"
          },
          "online": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexOption"
            }
          },
          "secrets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexSecret"
            }
          },
          "source": {
            "type": "string"
          },
          "supports_cpu": {
            "type": "boolean"
          },
          "supports_gpu": {
            "type": "boolean"
          },
          "update_available": {
            "type": "boolean",
            "description": "Whether `version` is newer than `installed_version`, decided by the\ndaemon.\n\nThe comparison is semver, and it belongs here rather than in each client\nfor the same reason `installed_version` does: the daemon is what reads\nthe installed manifest and owns the index, so it is the one place that\ncan answer without a client re-deriving it. A client that wants to\npresent the versions still has both.\n\n`false` when nothing is installed, when the installed version is at or\nahead of the index's, or when either version does not parse."
          },
          "version": {
            "type": "string"
          }
        }
      },
      "RegistryError": {
        "type": "object",
        "description": "The registry surface's error envelope, which carries one extra key.\n\nThese endpoints shipped before `error_code` existed, spelling the failure\nidentity as `error`. That key is still sent, because clients read it; the\nstandard `error_code` was added alongside rather than in place of it, so the\nwhole surface honors the \"`error_code` on every error\" rule without breaking\nanyone. Both name the same failure and always agree.\n\nDocumented as its own shape rather than folded into [`ErrorEnvelope`]:\n`error` appears *only* here, and putting it on the shared envelope would\ntell every other endpoint's reader to expect a key they will never receive.",
        "required": [
          "status",
          "error_code",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "The same identifier under the key this surface has always used. Retained\nfor clients written against it; prefer `error_code`.",
            "example": "not_found"
          },
          "error_code": {
            "type": "string",
            "description": "The stable identifier for this failure.",
            "example": "not_found"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail, when there is any to add."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
      "RegistryListResponse": {
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "backends"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistryBackend"
            }
          },
          "generated_at": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "SelectedAsset": {
        "type": "object",
        "required": [
          "target",
          "accel"
        ],
        "properties": {
          "accel": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Acceleration backends the selected build carries. A single entry is\nboth read and written as a bare string, a list of two or more as an\narray — a client that declares this field as a plain `String` still\nparses the catalog for every asset that carries one runtime."
          },
          "cuda_major": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "cuda_sm": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "cudnn": {
            "type": "boolean"
          },
          "target": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}